+++ tests/testthat/test-committee_meeting.R
@@ -1,0 +1,6 @@
+with_mock_dir("httptest2/commmeet", {
+  test_that("cong_committee_meeting works", {
+    x <- cong_committee_meeting()
+    expect_s3_class(x, 'tbl_df')
+  })
+})